Share sidebar action menu initializer across dashboard bundles
